// src/types/yahoo-finance.d.ts
export namespace YFModule {
// ...
  export interface EndDate3 {
    raw: number;
    fmt: string;
  }

  export interface Cash {
    raw: any;
    fmt: string;
    longFmt: string;
  }

  export interface ShortTermInvestments {
    raw: any;
    fmt: string;
    longFmt: string;
  }

  export interface NetReceivables {
    raw: any;
    fmt: string;
    longFmt: string;
  }

  export interface Inventory {
    raw: any;
    fmt: string;
    longFmt: string;
  }

  export interface OtherCurrentAssets {
    raw: any;
    fmt: string;
    longFmt: string;
  }

  export interface TotalCurrentAssets {
    raw: any;
    fmt: string;
    longFmt: string;
  }

  export interface LongTermInvestments {
    raw: any;
    fmt: string;
    longFmt: string;
  }

  export interface PropertyPlantEquipment {
    raw: any;
    fmt: string;
    longFmt: string;
  }

  export interface OtherAssets {
    raw: any;
    fmt: string;
    longFmt: string;
  }

  export interface TotalAssets3 {
    raw: any;
    fmt: string;
    longFmt: string;
  }

  export interface AccountsPayable {
    raw: any;
    fmt: string;
    longFmt: string;
  }

  export interface ShortLongTermDebt {
    raw: any;
    fmt: string;
    longFmt: string;
  }

  export interface OtherCurrentLiab {
    raw: any;
    fmt: string;
    longFmt: string;
  }

  export interface LongTermDebt {
    raw: any;
    fmt: string;
    longFmt: string;
  }

  export interface OtherLiab {
    raw: any;
    fmt: string;
    longFmt: string;
  }

  export interface TotalCurrentLiabilities {
    raw: any;
    fmt: string;
    longFmt: string;
  }

  export interface TotalLiab {
    raw: any;
    fmt: string;
    longFmt: string;
  }

  export interface CommonStock {
    raw: any;
    fmt: string;
    longFmt: string;
  }

  export interface RetainedEarnings {
    raw: any;
    fmt: string;
    longFmt: string;
  }

  export interface TreasuryStock {
    raw: number;
    fmt: string;
    longFmt: string;
  }

  export interface OtherStockholderEquity {
    raw: number;
    fmt: string;
    longFmt: string;
  }

  export interface TotalStockholderEquity {
    raw: any;
    fmt: string;
    longFmt: string;
  }

  export interface NetTangibleAssets {
    raw: any;
    fmt: string;
    longFmt: string;
  }

  export interface BalanceSheetStatement {
    maxAge: number;
    endDate: EndDate3;
    cash: Cash;
    shortTermInvestments: ShortTermInvestments;
    netReceivables: NetReceivables;
    inventory: Inventory;
    otherCurrentAssets: OtherCurrentAssets;
    totalCurrentAssets: TotalCurrentAssets;
    longTermInvestments: LongTermInvestments;
    propertyPlantEquipment: PropertyPlantEquipment;
    otherAssets: OtherAssets;
    totalAssets: TotalAssets3;
    accountsPayable: AccountsPayable;
    shortLongTermDebt: ShortLongTermDebt;
    otherCurrentLiab: OtherCurrentLiab;
    longTermDebt: LongTermDebt;
    otherLiab: OtherLiab;
    totalCurrentLiabilities: TotalCurrentLiabilities;
    totalLiab: TotalLiab;
    commonStock: CommonStock;
    retainedEarnings: RetainedEarnings;
    treasuryStock: TreasuryStock;
    otherStockholderEquity: OtherStockholderEquity;
    totalStockholderEquity: TotalStockholderEquity;
    netTangibleAssets: NetTangibleAssets;
  }
// ...
}